+++ app/workers/functions.php
@@ -328,12 +328,12 @@ class FunctionsV1 extends Worker
         
         Authorization::reset();
 
-        $runtime = (isset($runtimes[$function->getAttribute('env', '')]))
-            ? $runtimes[$function->getAttribute('env', '')]
+        $runtime = (isset($runtimes[$function->getAttribute('runtime', '')]))
+            ? $runtimes[$function->getAttribute('runtime', '')]
             : null;
 
         if(\is_null($runtime)) {
-            throw new Exception('Environment "'.$function->getAttribute('env', '').' is not supported');
+            throw new Exception('Runtime "'.$function->getAttribute('runtime', '').' is not supported');
         }
 
         $vars = \array_merge($function->getAttribute('vars', []), [